New Install protiuon of installer.

+++ install/pages/final.php
@@ -24,14 +24,90 @@ exponent_sessions_clearAllSessionData();
 
 global $user;
 
-if (unlink(BASE.'install/not_configured')) { 
+if (isset($_REQUEST['upgrade'])) { 
+// upgrades hit this
+    if (unlink(BASE.'install/not_configured')) { 
     
-    ?>
-    <h2><?php echo gt('You\'re all set!') ?></h2>
-<?php } else { ?>
-    <h2><?php echo gt('Hmmm..') ?></h2>
-    <p><?php echo gt('We weren\'t able to remove /install/not_configured. Remove this file manually to complete your installation.') ?></p>
-<?php } ?>
+            echo '<h2>' . gt('You\'re all set!') ."</h2>";
+            echo '<p>' . gt('Take me to your leader') ."</p>";
+
+        } else {
+
+            echo '<h2>' . gt('Hmmmm....') ."</h2>";
+            echo '<p>' . gt('We weren\'t able to remove /install/not_configured. Remove this file manually to complete your installation.') ."</p>";
+
+    } 
+
+} else {
+    
+    if (isset($_POST['username'])) {
+        $user = exponent_users_login($_POST['username'],$_POST['password']);
+
+        if (unlink(BASE.'install/not_configured')) { 
+            if ($user->id!=0) {
+                switch ($_POST['next']) {
+                    case 'migration':
+                        if (SEF_URLS) {
+                    	    header('Location: '.URL_FULL."migration/configure/");
+                        } else {
+                    	    header('Location: '.URL_FULL."index.php?module=migration&action=configure");
+                        }
+                        break;
+                    case 'configsite':
+                        if (SEF_URLS) {
+                    	    header('Location: '.URL_FULL."administration/configure_site/");
+                        } else {
+                    	    header('Location: '.URL_FULL."index.php?module=administration&action=configure_site");
+                        }
+                        break;
+                    default:
+                    	    header('Location: '.URL_FULL);
+                        break;
+                }
+            } else {
+                    
+                echo '<h2>' . gt('Hmmmm....') ."</h2>";
+                echo '<p>' . gt('We weren\'t able to remove /install/not_configured. Remove this file manually to complete your installation.') ."</p>";
+            
+            } 
+        }
+    }
+
+    echo '<h2>' . gt('You\'re all set!') ."</h2>";
+    echo '<p>' . gt('Log In and start managing your site') ."</p>";
+
+?>
+	<form action="index.php?page=final" method="POST">
+		<div class="text-control control ">
+			<label class="label">Username:</label><input type="text" class="text " size="25" value="" name="username">
+		</div>
+		<div class="password-control control ">
+			<label class="label">Password:</label><input type="password" class="password " size="25" value="" name="password">
+		</div>
+        <div class="formcontrol radiogroup">
+            <span class="label">And:</span>
+            <div class="formcontrol radiobutton">
+                <input type="radio" id="radiocontrol-1" class="radiobutton" value="migration" name="next">
+				<label for="radiocontrol-1">Take me to the Content Migration Wizard</label>
+            </div>
+            <div class="formcontrol radiobutton">
+                <input type="radio" id="radiocontrol-2" class="radiobutton" value="configsite" name="next">
+				<label for="radiocontrol-2">Take me to configure my site</label>
+            </div>
+            <div class="formcontrol radiobutton">
+                <input type="radio" id="radiocontrol-3" class="radiobutton" value="homepage" name="next">
+				<label for="radiocontrol-3">Just take me to my home page</label>
+            </div>
+        </div>
+		<div class="control buttongroup">
+			<button class="awesome large green">Log In</button>
+		</div>
+	</form>
+    
+<?php
+
+}
+?>
 
 <?php if (isset($_REQUEST['upgrade'])) { ?>
 <p><?php echo gt('Log back in to start using all your fancy new enhancements!') ?></p>
